+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Rename event OnDialogueStopped -> OnDialogueFinished
+
+### Removed
+- Package Gameplay System dependency
 
 ## [0.2.0] - 2025-11-14
 ### Added